Get quality assurance flow rate audit data for a primary quality assurance organization.

Get quality assurance flow rate audit data for a primary quality assurance organization.

get_qa_fra_in_pqao(bdate, edate, param, pqao)

Arguments

bdate
Beginning date to check. Year, month, day format.

edate
Ending date to check. Year, month, day format.

param
Pollutant parameter that site is measuring.

pqao
An encoding for a primary quality assurance organization. If unsure, use get_all_pqaos().

Value

API response containing operational information about the quality assurance flow rate audit data.

Examples

if (FALSE) {
bdate <- "20200101"
edate <- "20200131"
param <- "88101"
pqao <- "0013"
result <- get_qa_fra_in_pqao(bdate, edate, param, pqao)
result$Data
}